refactor wallets model
