+++ pyhodl/models/exchanges.py
@@ -26,7 +26,7 @@ from hal.streams.pretty_table import pretty_format_table
 from pyhodl.config import DATE_TIME_KEY, VALUE_KEY, NAN, \
     DEFAULT_FIAT
 from pyhodl.data.balance import parse_balance, save_balance
-from pyhodl.models.transactions import Wallet
+from pyhodl.models.wallets import Wallet
 from pyhodl.utils import datetime_to_str, get_delta_seconds, is_nan
